[SQUASH-1378] BUG FIXED : verify if the Automation Request list is empty before launching the  SQL deletion query

+++ tm/tm.service/src/main/java/org/squashtest/tm/service/internal/project/CustomGenericProjectManagerImpl.java
@@ -713,7 +713,9 @@ public class CustomGenericProjectManagerImpl implements CustomGenericProjectMana
 		List<Long> automationRequestIds = automationRequests.stream().map(AutomationRequest::getId).collect(Collectors.toList());
 
 		// delete remote automation request extenders
-		remoteAutomationRequestExtenderDao.deleteByAutomationRequestIds(automationRequestIds);
+		if(!automationRequestIds.isEmpty()){
+			remoteAutomationRequestExtenderDao.deleteByAutomationRequestIds(automationRequestIds);
+		}
 	}
 
 	@Override
